% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetRKAT.R
\docType{methods}
\name{applyRKAT,MetRKAT-method}
\alias{applyRKAT,MetRKAT-method}
\title{Apply the test to an already prepared method}
\usage{
\S4method{applyRKAT}{MetRKAT}(object, overwrite = FALSE, ...)
}
\arguments{
\item{MetRKAT}{}
}
\description{
Apply the test to an already prepared method
}

