% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetRKAT.R
\docType{methods}
\name{createIntervals,MetRKAT-method}
\alias{createIntervals,MetRKAT-method}
\title{Method to create the intervals based on the configuration options}
\usage{
\S4method{createIntervals}{MetRKAT}(object, overwrite = FALSE)
}
\arguments{
\item{MetRKAT}{}
}
\description{
Method to create the intervals based on the configuration options
}

