% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetRKAT.R
\docType{methods}
\name{initialize,MetRKAT-method}
\alias{initialize,MetRKAT-method}
\title{Constructor of MetRKAT class}
\usage{
\S4method{initialize}{MetRKAT}(.Object, ...)
}
\arguments{
\item{MetRKAT}{}
}
\description{
Constructor of MetRKAT class
}

