% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{MetRKAT-class}
\alias{MetRKAT-class}
\title{Main MetRKAT class}
\description{
Main MetRKAT class
}
\section{Slots}{

\describe{
\item{\code{data}}{data.frame. Contains CpG sites on rows and samples on columns }

\item{\code{wsize}}{numeric. The windows size}

\item{\code{wmethod}}{character. A character containing the method to build the windows}

\item{\code{gap}}{numeric. A non negative integer which defines how much the window will move in the iterative process}

\item{\code{intervals}}{data.frame. Contains the positions of the windows, and where they map to the annotation}

\item{\code{results}}{data.frame. Contains the result information. The first colums it's where is the window. Contains the distance method and the omnibus, if calculed, p-value.}

\item{\code{annotation}}{data.frame. Object containing the annotation of the CpGs}

\item{\code{distmethod}}{character. Indicate the distance method to compute the distances matrix}

\item{\code{max.na}}{numeric. Maximum proportion of NA values per row}
}}

