% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetRKAT.R
\name{MetRKAT}
\alias{MetRKAT}
\title{Main helper function to perform all the required steps for the analysis.}
\usage{
MetRKAT(y, data, wsize = 1000, wmethod = "default", gap = 0,
  distmethod = "euclidean", annotation = data.frame(), max.na = 0.5,
  dbname = NULL, ...)
}
\arguments{
\item{y}{Outcome variable. It can be either a dichotomous variable or a continous variable}
\item{data}{numerical data frame containing CpG sites on rows and samples on columns}
\item{wsize}{ The windows size}
\item{wmethod}{ A character containing the method to build the windows. Options are:

- default: It will use the wsize argument and the gap value to build the windows

- genes: The program will build the windows taking into account the annotation. It will collapse all the CpGs which have the same gene name in a window. If setting wmethod to genes, the function will ignore wsize and gap parameters.

- location_fixed: the function will generate windows taking into account wsize and gap parameters. When generating a window that overlaps another in the iterative process, it will be discarted.

- location: it's similar to location_fixed, but the duplicated windows will not be ignored.
}
\item{gap}{ a non negative integer which defines how much the window will move in the iterative process }

\item{distmethod}{a character indicating the distance method to compute the distances matrix. By default it is set to euclidean. If used with manhattan, the omnibus p-value will be calculated}

\item{annotation}{a data.frame object containing the annotation of the CpGs. If wmethod is set to 'genes' it's very important to note that the annotation data.frame must have a column called genes}

\item{max.na}{maximum proportion of NA values per row allowed before discarding it insteadof imputing them}

\item{dbname}{A character containing the name of the SQL database which will contain the analysis. By deafault set to NULL. If not NULL the analysis will run in a light way. Usefull for computers with low resources}

\item{...}{Arguments to be passed internally to MiRKAT functions. If your outcoume y is continous you should set out_type = continous.}
}
\description{
Main helper function to perform all the required steps for the analysis.
}

\examples{

#example <- MetRKAT(y = dicotomous_variable,
#data = methylation_data,
#wmethod = "location_fixed",
#wsize = 10000,
#gap = 1500,
#distmethod = c("euclidean", "manhattan"),
#max.na = 0.3)

### Accesing the results

#example@results
}


