% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetRKAT.R
\docType{methods}
\name{plotChromosome,MetRKAT-method}
\alias{plotChromosome,MetRKAT-method}
\title{Generates a chromosome ideogram plot}
\usage{
\S4method{plotChromosome}{MetRKAT}(object, chrom, pvals, cutoff = 0.05,
  logp = TRUE, hg_ideo = NULL, ...)
}
\arguments{
\item{object}{A MetRKAT object}
\item{chrom}{The chromosome name}
\item{pvals}{A charachter indicating which p-value sill be used: manhattan,euclidean or omnibus}
\item{cutoff}{A double which indicates the significance threshold}
\item{logp}{A logical value which indicates if -log10(p-value) is going to be used. By default set to TRUE}
\item{hg_ideo}{A Iviz ideogram. If nothing is giving, the function will use hg19 ideogram}


}
\description{
Generates a chromosome ideogram plot
}

\examples{
# plotChromosome(example, chrom = 'chr1', pvals = 'manhattan', cutoff = 0.05)
}
