% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetRKAT.R
\docType{methods}
\name{plotManhattan,MetRKAT-method}
\alias{plotManhattan,MetRKAT-method}
\title{Generates a manhattan plot}
\usage{
\S4method{plotManhattan}{MetRKAT}(object, pvals)
}
\arguments{
\item{object}{a MetRAT object}
\item{pvals}{a character indicating which p-values will be taken: euclidean, manhattan ot omnibus}
}
\description{
Generates a manhattan plot
}
\examples{
# plotManhattan(example, pvals = 'manhattan')
}

