% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetRKAT.R
\docType{methods}
\name{plotWindows,MetRKAT-method}
\alias{plotWindows,MetRKAT-method}
\title{Generates a continuous window plot}
\usage{
\S4method{plotWindows}{MetRKAT}(object, chrom, pvals, cutoff = 0.05,
  startpos = 0, endpos = Inf, logp = TRUE)
}
\arguments{
\item{object}{a MetRAT object}
\item{chrom}{the name od the choromosome to be ploted}
\item{pvals}{a character indicating which p-values will be taken: euclidean, manhattan ot omnibus}
\item{cutoff}{a double indicating the cutoff value}
\item{startpos}{The start position of the plot. By default = 0}
\item{endpos}{The en position to be ploted. By default set to Inf, which is the end of the chromosome}
\item{logp}{ A logical value which indicates if -log10(p-value) is going to be used. By default set to TRUE}

}
\description{
Generates a continuous window plot
}

\examples{
# plotWindows(example, chrom = 'chr1', pvals = 'euclidean')

}

