% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetRKAT.R
\docType{methods}
\name{toSQLite,MetRKAT-method}
\alias{toSQLite,MetRKAT-method}
\title{Transforms the object to use a SQLite database}
\usage{
\S4method{toSQLite}{MetRKAT}(object, dbname = NULL, overwrite = FALSE)
}
\arguments{
\item{object}{a MetRKAT object}
\item{dbname}{ a character variable indicating the name of the db}
\item{overwrite}{a logical argument. If there is a db created with that name, it indicates wether it should be overwrite. By default set to FALSE}
}
\description{
Transforms the object to use a SQLite database
}


